[Flex] Remove FlexLineBuilder and move code to RenderFlexibleBox. https://bugs.webkit.org/show_bug.cgi?id=292800 rdar://151030980
Reviewed by Alan Baradlay.

FlexLineBuilder is a very small class that really only has one main
function, computeNextFlexLine, which computes which flex items are
supposed to be on the line along with some other flexing data. Other
than that there are a couple of small helper functions and a
definition for FlexLayoutItem which is heavily used by
RenderFlexibleBox. Since all of this is fairly small and self-contained
code we can just have it live inside RenderFlexibleBox.

In order to accomplish this, the main changes are:
1.) Move the definition of FlexLayoutItem to RenderFlexibleBox.cpp
2.) Move many of the helper functions to RenderFlexibleBox. Some code
    is not needed anymore since they were just wrappers to functions in
    RenderFlexibleBox (e.g. FlexLineBuilder::isMultiLine)
3.) Plumb in a couple of arguments to computeNextLine. We can probably
    refactor this a bit more but this is a somewhat non-intrusive way in
    order to just accomplish the code moving.

Canonical link: https://commits.webkit.org/294755@main
