+++ Source/WebCore/SaferCPPExpectations/NoUncheckedPtrMemberCheckerExpectations
@@ -59,8 +59,6 @@ platform/mac/VideoPresentationInterfaceMac.mm
 platform/mac/WebPlaybackControlsManager.h
 platform/mock/DeviceOrientationClientMock.h
 rendering/BackgroundPainter.h
-rendering/BidiRun.h
-rendering/FlexLineBuilder.h
 rendering/GlyphDisplayListCache.cpp
 rendering/GridMasonryLayout.h
 rendering/GridTrackSizingAlgorithm.h
