+++ Source/WebCore/SaferCPPExpectations/UncheckedCallArgsCheckerExpectations
@@ -654,7 +654,6 @@ rendering/ContentfulPaintChecker.cpp
 rendering/CounterNode.cpp
 rendering/EllipsisBoxPainter.cpp
 rendering/FixedTableLayout.cpp
-rendering/FlexLineBuilder.cpp
 rendering/FloatingObjects.cpp
 rendering/GridMasonryLayout.cpp
 rendering/GridTrackSizingAlgorithm.cpp
