+++ Source/WebCore/SaferCPPExpectations/UncheckedLocalVarsCheckerExpectations
@@ -422,7 +422,6 @@ rendering/CaretRectComputation.cpp
 rendering/CounterNode.cpp
 rendering/EllipsisBoxPainter.cpp
 rendering/FixedTableLayout.cpp
-rendering/FlexLineBuilder.cpp
 rendering/FloatingObjects.cpp
 rendering/GridBaselineAlignment.cpp
 rendering/GridLayoutFunctions.cpp
